5.1.5 cleared the main.js from old code
